change: slackログイン時のcontrollerのAPIロジックをモデルに移動
